Offscreen rendering works now (This code is dirty)
Wanted to save this before I fix it up. I had two issues when writing this. One was that I made some shaders that didn't write to the buffer it seemed so I just straight up used the pbr shader and wrote its output to a texture. Then the rendering to a quad was pretty easy. I just needed a new shader for a 2d quad.