Deferred rendering (with multiple render targets)
An issue that popped up when I was doing this was that glFramebufferTexture2D kept getting passed one of the same texture id's twice, so the albedo map was getting the metalness and roughness texture which we didn't want (even though that's what I told the computer to do on accident). This code needs cleaned up, but it works. Next plan is to add diffuse irradiance, bloom, and add support for using the mat file's roughness and metalness values when no texture is specified (and then defaults for when those aren't specified either).