Object selection works now
Didn't use the stencil buffer because that only allows for selecting up to 255 objects (since 8 bit int texture). Instead I chose to just rerender the scene and output a similar calculation to the position texture. This way I can use 255^4 objects theoretically. I assign an integer id to each game object (generating this id could be much more robust, but this is just to get it working) and then I use that value to create a color which I store in a texture, then I just read the pixels back from that offscreen buffer, turn that color back into an id, and then use that to set an isSelected variable on the object with that id.